Refactor code to support different teachers in the future
