--- src/lightly_train/_modules/dinov2/layers/__init__.py
@@ -1,22 +1,0 @@
-#
-# Copyright (c) Lightly AG and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the license found in the
-# LICENSE file in the root directory of this source tree.
-#
-
-from lightly_train._modules.dinov2.layers.attention import MemEffAttention
-from lightly_train._modules.dinov2.layers.block import NestedTensorBlock
-from lightly_train._modules.dinov2.layers.mlp import Mlp
-from lightly_train._modules.dinov2.layers.patch_embed import PatchEmbed
-from lightly_train._modules.dinov2.layers.swiglu_ffn import SwiGLUFFN, SwiGLUFFNFused
-
-__all__ = [
-    "MemEffAttention",
-    "NestedTensorBlock",
-    "Mlp",
-    "PatchEmbed",
-    "SwiGLUFFN",
-    "SwiGLUFFNFused",
-]
