+++ src/lightly_train/_modules/teachers/__init__.py
@@ -1,0 +1,10 @@
+#
+# Copyright (c) Lightly AG and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the license found in the
+# LICENSE file in the root directory of this source tree.
+#
+from lightly_train._modules.teachers.build_teacher import get_teacher
+
+__all__ = ["get_teacher"]
