+++ src/lightly_train/_modules/teachers/dinov2/__init__.py
@@ -8,6 +8,6 @@
 
 __version__ = "0.0.1"
 
-from lightly_train._modules.dinov2.build_teacher import get_teacher_model
+from lightly_train._modules.teachers.dinov2.build_teacher import get_dinov2_teacher
 
-__all__ = ["get_teacher_model"]
+__all__ = ["get_dinov2_teacher"]
