+++ src/lightly_train/_modules/teachers/dinov2/configs/__init__.py
