+++ src/lightly_train/_modules/teachers/dinov2/configs/eval/vitb14_pretrain.yaml
