+++ src/lightly_train/_modules/teachers/dinov2/configs/eval/vitg14_pretrain.yaml
