+++ src/lightly_train/_modules/teachers/dinov2/configs/eval/vitl14_pretrain.yaml
