+++ src/lightly_train/_modules/teachers/dinov2/configs/eval/vits14_pretrain.yaml
