+++ src/lightly_train/_modules/teachers/dinov2/configs/eval/vits14_reg4_pretrain.yaml
