+++ src/lightly_train/_modules/teachers/dinov2/configs/ssl_default_config.yaml
