+++ src/lightly_train/_modules/teachers/dinov2/layers/layer_scale.py
