+++ src/lightly_train/_modules/teachers/dinov2/layers/patch_embed.py
