+++ src/lightly_train/_modules/teachers/dinov2/layers/swiglu_ffn.py
